refactor: moved all dialogs on `workspace/governance/customers` tosheets
